refactor: api.ts
